Moves files from settings to settings/general
